pyproject.toml
[tool.poetry]
name = "AgePrediction-UTKFace"
version = "0.1.0"
description = "Age Prediction on the UTK Face Dataset"
authors = ["Moritz Mistol <moritz.mistol@gmail.com>"]
readme = "README.md"
documentation = "https://MoritzM00.github.io/AgePrediction-UTKFace"
repository = "https://github.com/MoritzM00/AgePrediction-UTKFace"
license = "MIT"
keywords = ["Python", "Poetry"]
packages = [
    { include = "src" },
]
include = [
    "src/conf/*"
]

[tool.poetry.dependencies]
python = "^3.9"
jupyter = "^1.0.0"
numpy = "^1.24.3"
hydra-core = "^1.3.2"
pandas = "^2.0.1"
matplotlib = "^3.7.1"
wandb = "^0.15.0"
scikit-learn = "^1.2.2"
seaborn = "^0.12.2"
protobuf = "<3.21"

[tool.poetry.scripts]
train = 'src.scripts.main:main'


[tool.poetry.group.dev.dependencies]
pytest = ">=7.2.2"
pre-commit = ">=3.1.1"
pdoc3 = ">=0.10.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
